+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/CruiseControlTest.java
@@ -43,6 +43,7 @@ import io.strimzi.api.kafka.model.KafkaBuilder;
 import io.strimzi.api.kafka.model.KafkaResources;
 import io.strimzi.api.kafka.model.MetricsConfig;
 import io.strimzi.api.kafka.model.SystemPropertyBuilder;
+import io.strimzi.api.kafka.model.balancing.BrokerCapacityBuilder;
 import io.strimzi.api.kafka.model.storage.EphemeralStorage;
 import io.strimzi.api.kafka.model.storage.JbodStorage;
 import io.strimzi.api.kafka.model.storage.JbodStorageBuilder;
@@ -66,6 +67,7 @@ import io.strimzi.test.annotations.ParallelSuite;
 import io.strimzi.test.annotations.ParallelTest;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.AfterAll;
 
 import java.util.ArrayList;
@@ -110,15 +112,19 @@ public class CruiseControlTest {
 
     private static final SharedEnvironmentProvider SHARED_ENV_PROVIDER = new MockSharedEnvironmentProvider();
 
-    private final String namespace = "test";
-    private final String cluster = "foo";
-    private final int replicas = 1;
-    private final String image = "my-image:latest";
-    private final int healthDelay = 120;
-    private final int healthTimeout = 30;
-    private final String minInsyncReplicas = "2";
+    private static final String NAMESPACE = "test";
+    private static final String CLUSTER = "foo";
+    private static final int REPLICAS = 1;
+    private static final String IMAGE = "my-image:latest";
+    private static final int HEALTH_DELAY = 120;
+    private static final int HEALTH_TIMEOUT = 30;
+    private static final String MIN_INSYNC_REPLICAS = "2";
+    private static final String BROKER_CAPACITY_CPU = "6.0";
+    private static final String BROKER_CAPACITY_OVERRIDE_CPU = "2.0";
+    private static final String RESOURCE_LIMIT_CPU = "3.0";
+    private static final String RESOURCE_REQUESTS_CPU = "4.0";
 
-    private final Map<String, Object> kafkaConfig = singletonMap(CruiseControl.MIN_INSYNC_REPLICAS, minInsyncReplicas);
+    private final Map<String, Object> kafkaConfig = singletonMap(CruiseControl.MIN_INSYNC_REPLICAS, MIN_INSYNC_REPLICAS);
     private final Map<String, Object> ccConfig = new HashMap<>() {{
             putAll(CruiseControlConfiguration.getCruiseControlDefaultPropertiesMap());
             put("num.partition.metrics.windows", "2");
@@ -150,14 +156,14 @@ public class CruiseControlTest {
     private final Kafka kafka = createKafka(cruiseControlSpec);
     private final CruiseControl cc = createCruiseControl(kafka);
 
-    private Map<String, Storage> createStorageMap(Kafka kafkaAssembly)    {
+    private Map<String, Storage> createStorageMap(Kafka kafkaAssembly) {
         Map<String, Storage> storage = new HashMap<>();
         storage.put("kafka", kafkaAssembly.getSpec().getKafka().getStorage());
 
         return storage;
     }
 
-    private Map<String, ResourceRequirements> createResourceRequirementsMap(Kafka kafkaAssembly)    {
+    private Map<String, ResourceRequirements> createResourceRequirementsMap(Kafka kafkaAssembly) {
         Map<String, ResourceRequirements> resources = new HashMap<>();
         resources.put("kafka", kafkaAssembly.getSpec().getKafka().getResources());
 
@@ -178,43 +184,43 @@ public class CruiseControlTest {
     }
 
     private Kafka createKafka(CruiseControlSpec cruiseControlSpec) {
-        return new KafkaBuilder(ResourceUtils.createKafka(namespace, cluster, replicas, image, healthDelay, healthTimeout))
+        return new KafkaBuilder(ResourceUtils.createKafka(NAMESPACE, CLUSTER, REPLICAS, IMAGE, HEALTH_DELAY, HEALTH_TIMEOUT))
                 .editSpec()
                     .editKafka()
                         .withConfig(kafkaConfig)
                         .withStorage(kafkaStorage)
-                   .endKafka()
-                  .withCruiseControl(cruiseControlSpec)
+                    .endKafka()
+                    .withCruiseControl(cruiseControlSpec)
                 .endSpec()
                 .build();
     }
 
-    private Map<String, String> expectedLabels(String name)    {
-        return TestUtils.map(Labels.STRIMZI_CLUSTER_LABEL, this.cluster,
+    private Map<String, String> expectedLabels(String name) {
+        return TestUtils.map(Labels.STRIMZI_CLUSTER_LABEL, this.CLUSTER,
                 "my-user-label", "cromulent",
                 Labels.STRIMZI_KIND_LABEL, Kafka.RESOURCE_KIND,
                 Labels.STRIMZI_NAME_LABEL, name,
                 Labels.STRIMZI_COMPONENT_TYPE_LABEL, CruiseControl.COMPONENT_TYPE,
                 Labels.KUBERNETES_NAME_LABEL, CruiseControl.COMPONENT_TYPE,
-                Labels.KUBERNETES_INSTANCE_LABEL, this.cluster,
-                Labels.KUBERNETES_PART_OF_LABEL, Labels.APPLICATION_NAME + "-" + this.cluster,
+                Labels.KUBERNETES_INSTANCE_LABEL, this.CLUSTER,
+                Labels.KUBERNETES_PART_OF_LABEL, Labels.APPLICATION_NAME + "-" + this.CLUSTER,
                 Labels.KUBERNETES_MANAGED_BY_LABEL, AbstractModel.STRIMZI_CLUSTER_OPERATOR_NAME);
     }
 
-    private Map<String, String> expectedSelectorLabels()    {
+    private Map<String, String> expectedSelectorLabels() {
         return Labels.fromMap(expectedLabels()).strimziSelectorLabels().toMap();
     }
 
-    private Map<String, String> expectedLabels()    {
-        return expectedLabels(CruiseControlResources.deploymentName(cluster));
+    private Map<String, String> expectedLabels() {
+        return expectedLabels(CruiseControlResources.deploymentName(CLUSTER));
     }
 
     private List<EnvVar> getExpectedEnvVars() {
         List<EnvVar> expected = new ArrayList<>();
         expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_CRUISE_CONTROL_METRICS_ENABLED).withValue(Boolean.toString(CruiseControl.DEFAULT_CRUISE_CONTROL_METRICS_ENABLED)).build());
-        expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_STRIMZI_KAFKA_BOOTSTRAP_SERVERS).withValue(KafkaResources.bootstrapServiceName(cluster) + ":" + KafkaCluster.REPLICATION_PORT).build());
+        expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_STRIMZI_KAFKA_BOOTSTRAP_SERVERS).withValue(KafkaResources.bootstrapServiceName(CLUSTER) + ":" + KafkaCluster.REPLICATION_PORT).build());
         expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_STRIMZI_KAFKA_GC_LOG_ENABLED).withValue(Boolean.toString(JvmOptions.DEFAULT_GC_LOGGING_ENABLED)).build());
-        expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_MIN_INSYNC_REPLICAS).withValue(minInsyncReplicas).build());
+        expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_MIN_INSYNC_REPLICAS).withValue(MIN_INSYNC_REPLICAS).build());
         expected.add(new EnvVarBuilder().withName(ENV_VAR_CRUISE_CONTROL_CAPACITY_CONFIGURATION).withValue(cc.capacity.toString()).build());
         expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_API_SSL_ENABLED).withValue(Boolean.toString(CruiseControlConfigurationParameters.DEFAULT_WEBSERVER_SSL_ENABLED)).build());
         expected.add(new EnvVarBuilder().withName(CruiseControl.ENV_VAR_API_AUTH_ENABLED).withValue(Boolean.toString(CruiseControlConfigurationParameters.DEFAULT_WEBSERVER_SECURITY_ENABLED)).build());
@@ -240,6 +246,17 @@ public class CruiseControlTest {
 
     private static boolean isJBOD(Map<String, Object> brokerCapacity) {
         return brokerCapacity.get(Capacity.DISK_KEY) instanceof JsonObject;
+    }
+
+    public Kafka kafkaSpec(CruiseControlSpec cruiseControlSpec, ResourceRequirements resourceRequirements) {
+        return new KafkaBuilder()
+                .withNewSpec()
+                    .withNewKafka()
+                        .withResources(resourceRequirements)
+                    .endKafka()
+                    .withCruiseControl(cruiseControlSpec)
+                .endSpec()
+                .build();
     }
 
     @ParallelTest
@@ -273,7 +290,7 @@ public class CruiseControlTest {
         Map<String, Quantity> requests = Map.of(Capacity.RESOURCE_TYPE, new Quantity("400m"));
         Map<String, Quantity> limits = Map.of(Capacity.RESOURCE_TYPE, new Quantity("0.5"));
 
-        resource = new KafkaBuilder(ResourceUtils.createKafka(namespace, cluster, replicas, image, healthDelay, healthTimeout))
+        resource = new KafkaBuilder(ResourceUtils.createKafka(NAMESPACE, CLUSTER, REPLICAS, IMAGE, HEALTH_DELAY, HEALTH_TIMEOUT))
             .editSpec()
                 .editKafka()
                     .withVersion(KafkaVersionTestUtils.DEFAULT_KAFKA_VERSION)
@@ -283,7 +300,7 @@ public class CruiseControlTest {
                 .withCruiseControl(cruiseControlSpec)
             .endSpec()
             .build();
-        
+
         capacity = new Capacity(Reconciliation.DUMMY_RECONCILIATION, resource.getSpec(), NODES, createStorageMap(resource), createResourceRequirementsMap(resource));
         String cpuCapacity = new CpuCapacity(userDefinedCpuCapacity).toString();
 
@@ -339,16 +356,16 @@ public class CruiseControlTest {
 
         TreeMap<Integer, BrokerCapacity> capacityEntries = capacity.getCapacityEntries();
 
-        assertThat(capacityEntries.get(broker0).getCpu().toString(), is(new CpuCapacity(userDefinedCpuCapacity).toString()));
+        assertThat(capacityEntries.get(broker0).getCpu().toString(), is(new CpuCapacity(userDefinedCpuCapacityOverride0).toString()));
         assertThat(capacityEntries.get(broker0).getInboundNetwork(), is(Capacity.getThroughputInKiB(inboundNetworkOverride0)));
         assertThat(capacityEntries.get(broker0).getOutboundNetwork(), is(BrokerCapacity.DEFAULT_OUTBOUND_NETWORK_CAPACITY_IN_KIB_PER_SECOND));
 
         // When the same broker id is specified in brokers list of multiple overrides, use the value specified in the first override.
-        assertThat(capacityEntries.get(broker1).getCpu().toString(), is(new CpuCapacity(userDefinedCpuCapacity).toString()));
+        assertThat(capacityEntries.get(broker1).getCpu().toString(), is(new CpuCapacity(userDefinedCpuCapacityOverride0).toString()));
         assertThat(capacityEntries.get(broker1).getInboundNetwork(), is(Capacity.getThroughputInKiB(inboundNetworkOverride0)));
         assertThat(capacityEntries.get(broker1).getOutboundNetwork(), is(BrokerCapacity.DEFAULT_OUTBOUND_NETWORK_CAPACITY_IN_KIB_PER_SECOND));
 
-        assertThat(capacityEntries.get(broker2).getCpu().toString(), is(new CpuCapacity(userDefinedCpuCapacity).toString()));
+        assertThat(capacityEntries.get(broker2).getCpu().toString(), is(new CpuCapacity(userDefinedCpuCapacityOverride0).toString()));
         assertThat(capacityEntries.get(broker2).getInboundNetwork(), is(Capacity.getThroughputInKiB(inboundNetworkOverride0)));
 
         assertThat(getCapacityConfigurationFromEnvVar(resource, ENV_VAR_CRUISE_CONTROL_CAPACITY_CONFIGURATION), is(capacity.toString()));
@@ -359,7 +376,7 @@ public class CruiseControlTest {
         requests = Map.of(Capacity.RESOURCE_TYPE, new Quantity(userDefinedCpuCapacity));
         limits = Map.of(Capacity.RESOURCE_TYPE, new Quantity("0.5"));
 
-        resource = new KafkaBuilder(ResourceUtils.createKafka(namespace, cluster, replicas, image, healthDelay, healthTimeout))
+        resource = new KafkaBuilder(ResourceUtils.createKafka(NAMESPACE, CLUSTER, REPLICAS, IMAGE, HEALTH_DELAY, HEALTH_TIMEOUT))
             .editSpec()
                 .editKafka()
                 .withVersion(KafkaVersionTestUtils.DEFAULT_KAFKA_VERSION)
@@ -371,7 +388,7 @@ public class CruiseControlTest {
             .build();
 
         capacity = new Capacity(Reconciliation.DUMMY_RECONCILIATION, resource.getSpec(), NODES, createStorageMap(resource), createResourceRequirementsMap(resource));
-        cpuCapacity = new CpuCapacity(userDefinedCpuCapacity).toString();
+        cpuCapacity = new CpuCapacity(userDefinedCpuCapacityOverride0).toString();
 
         brokerEntries = capacity.generateCapacityConfig().getJsonArray(Capacity.CAPACITIES_KEY);
         for (Object brokerEntry : brokerEntries) {
@@ -382,7 +399,7 @@ public class CruiseControlTest {
 
     @ParallelTest
     public void testBrokerCapacitiesWithPools() {
-        Kafka kafka = new KafkaBuilder(ResourceUtils.createKafka(namespace, cluster, replicas, image, healthDelay, healthTimeout))
+        Kafka kafka = new KafkaBuilder(ResourceUtils.createKafka(NAMESPACE, CLUSTER, REPLICAS, IMAGE, HEALTH_DELAY, HEALTH_TIMEOUT))
                 .editSpec()
                     .withCruiseControl(cruiseControlSpec)
                 .endSpec()
@@ -464,8 +481,8 @@ public class CruiseControlTest {
 
     @ParallelTest
     public void testFromConfigMap() {
-        assertThat(cc.namespace, is(namespace));
-        assertThat(cc.cluster, is(cluster));
+        assertThat(cc.namespace, is(NAMESPACE));
+        assertThat(cc.cluster, is(CLUSTER));
         assertThat(cc.getImage(), is(ccImage));
     }
 
@@ -477,8 +494,8 @@ public class CruiseControlTest {
 
         assertThat(containers.size(), is(1));
 
-        assertThat(dep.getMetadata().getName(), is(CruiseControlResources.deploymentName(cluster)));
-        assertThat(dep.getMetadata().getNamespace(), is(namespace));
+        assertThat(dep.getMetadata().getName(), is(CruiseControlResources.deploymentName(CLUSTER)));
+        assertThat(dep.getMetadata().getNamespace(), is(NAMESPACE));
         TestUtils.checkOwnerReference(dep, kafka);
 
         // checks on the main Cruise Control container
@@ -501,19 +518,19 @@ public class CruiseControlTest {
 
         Volume volume = volumes.stream().filter(vol -> CruiseControl.TLS_CC_CERTS_VOLUME_NAME.equals(vol.getName())).findFirst().orElseThrow();
         assertThat(volume, is(notNullValue()));
-        assertThat(volume.getSecret().getSecretName(), is(CruiseControlResources.secretName(cluster)));
+        assertThat(volume.getSecret().getSecretName(), is(CruiseControlResources.secretName(CLUSTER)));
 
         volume = volumes.stream().filter(vol -> CruiseControl.TLS_CA_CERTS_VOLUME_NAME.equals(vol.getName())).findFirst().orElseThrow();
         assertThat(volume, is(notNullValue()));
-        assertThat(volume.getSecret().getSecretName(), is(AbstractModel.clusterCaCertSecretName(cluster)));
+        assertThat(volume.getSecret().getSecretName(), is(AbstractModel.clusterCaCertSecretName(CLUSTER)));
 
         volume = volumes.stream().filter(vol -> CruiseControl.LOG_AND_METRICS_CONFIG_VOLUME_NAME.equals(vol.getName())).findFirst().orElseThrow();
         assertThat(volume, is(notNullValue()));
-        assertThat(volume.getConfigMap().getName(), is(CruiseControlResources.logAndMetricsConfigMapName(cluster)));
+        assertThat(volume.getConfigMap().getName(), is(CruiseControlResources.logAndMetricsConfigMapName(CLUSTER)));
 
         volume = volumes.stream().filter(vol -> CruiseControl.API_AUTH_CONFIG_VOLUME_NAME.equals(vol.getName())).findFirst().orElseThrow();
         assertThat(volume, is(notNullValue()));
-        assertThat(volume.getSecret().getSecretName(), is(CruiseControlResources.apiSecretName(cluster)));
+        assertThat(volume.getSecret().getSecretName(), is(CruiseControlResources.apiSecretName(CLUSTER)));
 
         volume = volumes.stream().filter(vol -> VolumeUtils.STRIMZI_TMP_DIRECTORY_DEFAULT_VOLUME_NAME.equals(vol.getName())).findFirst().orElseThrow();
         assertThat(volume, is(notNullValue()));
@@ -546,7 +563,7 @@ public class CruiseControlTest {
     }
 
     @ParallelTest
-    public void testEnvVars()   {
+    public void testEnvVars() {
         assertThat(cc.getEnvVars(), is(getExpectedEnvVars()));
     }
 
@@ -581,7 +598,7 @@ public class CruiseControlTest {
                 .withImage(ccImage)
                 .withNewTemplate()
                     .withNewCruiseControlContainer()
-                       .withEnv(envVar1, envVar2)
+                        .withEnv(envVar1, envVar2)
                     .endCruiseControlContainer()
                 .endTemplate()
                 .build();
@@ -632,11 +649,11 @@ public class CruiseControlTest {
     }
 
     @ParallelTest
-    public void testGenerateService()   {
+    public void testGenerateService() {
         Service svc = cc.generateService();
 
         assertThat(svc.getSpec().getType(), is("ClusterIP"));
-        assertThat(svc.getMetadata().getLabels(), is(expectedLabels(CruiseControlResources.serviceName(cluster))));
+        assertThat(svc.getMetadata().getLabels(), is(expectedLabels(CruiseControlResources.serviceName(CLUSTER))));
         assertThat(svc.getSpec().getSelector(), is(expectedSelectorLabels()));
         assertThat(svc.getSpec().getPorts().size(), is(1));
         assertThat(svc.getSpec().getPorts().get(0).getName(), is(CruiseControl.REST_API_PORT_NAME));
@@ -752,7 +769,7 @@ public class CruiseControlTest {
         svcLabels.putAll(expectedLabels());
         Service svc = cc.generateService();
         assertThat(svc.getMetadata().getLabels(), is(svcLabels));
-        assertThat(svc.getMetadata().getAnnotations(),  is(svcAnots));
+        assertThat(svc.getMetadata().getAnnotations(), is(svcAnots));
         assertThat(svc.getSpec().getIpFamilyPolicy(), is("PreferDualStack"));
         assertThat(svc.getSpec().getIpFamilies(), contains("IPv6", "IPv4"));
 
@@ -786,6 +803,94 @@ public class CruiseControlTest {
 
         assertThat(ccContainer.getResources().getLimits(), is(limits));
         assertThat(ccContainer.getResources().getRequests(), is(requests));
+    }
+
+    @ParallelTest
+    public void testCpuCapacityGeneration() {
+        Set<NodeRef> nodes = Set.of(
+                new NodeRef("foo-pool1-0", 0, "pool1", false, true),
+                new NodeRef("foo-pool1-1", 1, "pool1", false, true),
+                new NodeRef("foo-pool1-2", 2, "pool1", false, true)
+        );
+
+        Map<String, Storage> storage = new HashMap<>();
+        storage.put("pool1", new JbodStorageBuilder().withVolumes(new PersistentClaimStorageBuilder().withId(0).withSize("100Gi").build()).build());
+
+        /* In this test case, when override is set for a broker, it takes precedence over the general broker capacity setting for that
+           specific broker, but the general broker capacity takes precedence over the Kafka resource request */
+        ResourceRequirements resourceRequirements = new ResourceRequirementsBuilder()
+                .addToLimits("cpu", new Quantity(RESOURCE_LIMIT_CPU))
+                .addToRequests("cpu", new Quantity(RESOURCE_REQUESTS_CPU))
+                .build();
+
+        io.strimzi.api.kafka.model.balancing.BrokerCapacity brokerCapacityOne = new BrokerCapacityBuilder()
+                .withCpu(BROKER_CAPACITY_CPU)
+                .withOverrides()
+                    .addNewOverride()
+                        .addToBrokers(0)
+                        .withCpu(BROKER_CAPACITY_OVERRIDE_CPU)
+                    .endOverride()
+                .build();
+
+        verifyBrokerCapacity(nodes, storage, resourceRequirements, brokerCapacityOne, BROKER_CAPACITY_OVERRIDE_CPU, BROKER_CAPACITY_CPU, BROKER_CAPACITY_CPU);
+
+        // In this test case, when override is set for a broker, it takes precedence over the Kafka resource request
+        io.strimzi.api.kafka.model.balancing.BrokerCapacity brokerCapacityTwo = new BrokerCapacityBuilder()
+                .withOverrides()
+                    .addNewOverride()
+                        .addToBrokers(0)
+                        .withCpu(BROKER_CAPACITY_OVERRIDE_CPU)
+                    .endOverride()
+                .build();
+
+        verifyBrokerCapacity(nodes, storage, resourceRequirements, brokerCapacityTwo, BROKER_CAPACITY_OVERRIDE_CPU, RESOURCE_REQUESTS_CPU, RESOURCE_REQUESTS_CPU);
+
+        /* In this test case, when neither the override nor the CPU resource request are configured but the CPU
+           resource limit for CPU is set for the Kafka brokers; therefore, resource limit will be used as the CPU capacity */
+        ResourceRequirements resourceRequirements3 = new ResourceRequirementsBuilder()
+                .addToLimits("cpu", new Quantity(RESOURCE_LIMIT_CPU))
+                .build();
+
+        io.strimzi.api.kafka.model.balancing.BrokerCapacity brokerCapacityThree = new BrokerCapacityBuilder()
+                .build();
+
+        verifyBrokerCapacity(nodes, storage, resourceRequirements3, brokerCapacityThree, RESOURCE_LIMIT_CPU, RESOURCE_LIMIT_CPU, RESOURCE_LIMIT_CPU);
+
+        /* In this test case, when neither the override nor the Kafka resource requests or limits for CPU are configured,
+           the CPU capacity will be set to DEFAULT_CPU_CORE_CAPACITY */
+        ResourceRequirements resourceRequirements2 = new ResourceRequirementsBuilder()
+                .build();
+
+        verifyBrokerCapacity(nodes, storage, resourceRequirements2, brokerCapacityThree, BrokerCapacity.DEFAULT_CPU_CORE_CAPACITY, BrokerCapacity.DEFAULT_CPU_CORE_CAPACITY, BrokerCapacity.DEFAULT_CPU_CORE_CAPACITY);
+    }
+
+    private void verifyBrokerCapacity(Set<NodeRef> nodes,
+                                     Map<String, Storage> storage,
+                                     ResourceRequirements resourceRequirements,
+                                     io.strimzi.api.kafka.model.balancing.BrokerCapacity brokerCapacity,
+                                     String brokerOneCpuValue,
+                                     String brokerTwoCpuValue,
+                                     String brokerThreeCpuValue) {
+
+        HashMap<String, ResourceRequirements> resources = new HashMap<>();
+        resources.put("pool1", resourceRequirements);
+
+        CruiseControlSpec cruiseControlSpec = new CruiseControlSpecBuilder()
+                .withBrokerCapacity(brokerCapacity)
+                .build();
+
+        CruiseControl cc = CruiseControl.fromCrd(Reconciliation.DUMMY_RECONCILIATION, kafkaSpec(cruiseControlSpec, resourceRequirements), VERSIONS, nodes, storage, resources, SHARED_ENV_PROVIDER);
+        Capacity capacity = cc.getCapacity();
+
+        TreeMap<Integer, BrokerCapacity> map = capacity.getCapacityEntries();
+        BrokerCapacity broker0 = map.get(0);
+        BrokerCapacity broker1 = map.get(1);
+        BrokerCapacity broker2 = map.get(2);
+
+        assertThat(capacity, is(notNullValue()));
+        assertThat(broker0.getCpu().getCores(), is(Matchers.equalTo(brokerOneCpuValue)));
+        assertThat(broker1.getCpu().getCores(), is(Matchers.equalTo(brokerTwoCpuValue)));
+        assertThat(broker2.getCpu().getCores(), is(Matchers.equalTo(brokerThreeCpuValue)));
     }
 
     @ParallelTest
@@ -834,12 +939,12 @@ public class CruiseControlTest {
         CruiseControlSpec cruiseControlSpec = new CruiseControlSpecBuilder()
                 .withImage(ccImage)
                 .withNewLivenessProbe()
-                        .withInitialDelaySeconds(healthDelay)
-                        .withTimeoutSeconds(healthTimeout)
+                        .withInitialDelaySeconds(HEALTH_DELAY)
+                        .withTimeoutSeconds(HEALTH_TIMEOUT)
                 .endLivenessProbe()
                 .withNewReadinessProbe()
-                       .withInitialDelaySeconds(healthDelay)
-                       .withTimeoutSeconds(healthTimeout)
+                       .withInitialDelaySeconds(HEALTH_DELAY)
+                       .withTimeoutSeconds(HEALTH_TIMEOUT)
                 .endReadinessProbe()
                 .build();
 
@@ -852,10 +957,10 @@ public class CruiseControlTest {
         // checks on the main Cruise Control container
         Container ccContainer = containers.stream().filter(container -> ccImage.equals(container.getImage())).findFirst().orElseThrow();
         assertThat(ccContainer.getImage(), is(cc.image));
-        assertThat(ccContainer.getLivenessProbe().getInitialDelaySeconds(), is(healthDelay));
-        assertThat(ccContainer.getLivenessProbe().getTimeoutSeconds(), is(healthTimeout));
-        assertThat(ccContainer.getReadinessProbe().getInitialDelaySeconds(), is(healthDelay));
-        assertThat(ccContainer.getReadinessProbe().getTimeoutSeconds(), is(healthTimeout));
+        assertThat(ccContainer.getLivenessProbe().getInitialDelaySeconds(), is(HEALTH_DELAY));
+        assertThat(ccContainer.getLivenessProbe().getTimeoutSeconds(), is(HEALTH_TIMEOUT));
+        assertThat(ccContainer.getReadinessProbe().getInitialDelaySeconds(), is(HEALTH_DELAY));
+        assertThat(ccContainer.getReadinessProbe().getTimeoutSeconds(), is(HEALTH_TIMEOUT));
     }
 
     @ParallelTest
@@ -994,7 +1099,7 @@ public class CruiseControlTest {
                 .endPodSelector()
                 .build();
 
-        NetworkPolicy np = cc.generateNetworkPolicy(namespace, null);
+        NetworkPolicy np = cc.generateNetworkPolicy(NAMESPACE, null);
 
         assertThat(np.getSpec().getIngress().stream().filter(ing -> ing.getPorts().get(0).getPort().equals(new IntOrString(CruiseControl.REST_API_PORT))).findFirst().orElse(null), is(notNullValue()));